requirements.txt
asgiref==3.8.1
Authlib==1.3.2
bcrypt==3.2.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
cryptography==44.0.0
cssselect==1.2.0
cssutils==2.11.1
ecdsa==0.19.0
emails==0.6
fastapi==0.68.1
h11==0.14.0
idna==3.10
lxml==5.3.0
more-itertools==10.5.0
passlib==1.7.4
premailer==3.10.0
pyasn1==0.6.1
pycparser==2.22
pydantic==1.8.2
python-dateutil==2.9.0.post0
python-dotenv==0.19.0
python-jose==3.3.0
python-multipart==0.0.5
requests==2.32.3
rsa==4.9
six==1.16.0
starlette==0.14.2
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.15.0
validators==0.34.0
weaviate-client==3.24.1
